refactor: extract `Paragraph` component
